+++ internal/provider/provider.go
@@ -241,6 +241,7 @@ func (p *SdwanProvider) Configure(ctx context.Context, req provider.ConfigureReq
 
 func (p *SdwanProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
+		NewArchiveFeatureTemplateResource,
 		NewCEdgeAAAFeatureTemplateResource,
 		NewCEdgeGlobalFeatureTemplateResource,
 		NewCEdgeIGMPFeatureTemplateResource,
@@ -361,6 +362,7 @@ func (p *SdwanProvider) Resources(ctx context.Context) []func() resource.Resourc
 
 func (p *SdwanProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewArchiveFeatureTemplateDataSource,
 		NewCEdgeAAAFeatureTemplateDataSource,
 		NewCEdgeGlobalFeatureTemplateDataSource,
 		NewCEdgeIGMPFeatureTemplateDataSource,