+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add `sdwan_mode_feature_template` resource and data source
 - Add `sdwan_multicast_feature_template` resource and data source
 - Add `sdwan_ntp_feature_template` resource and data source
+- Add `sdwan_archive_feature_template` resource and data source
 
 ## 0.3.6